Fixed errors with types

+++ src/NashEquilibrium/GeneralSum1.java
@@ -4,7 +4,7 @@ import play.NormalFormGame;
 
 public class GeneralSum1 {
     public static void main(String[] args) {
-        int[][] test1 = new int[3][3];
+        double[][] test1 = new double[3][3];
         test1[0][0] = 0;
         test1[0][1] = 0;
         test1[0][2] = 1;
@@ -15,7 +15,7 @@ public class GeneralSum1 {
         test1[2][1] = 0;
         test1[2][2] = 0;
 
-        int[][] test2 = new int[3][3];
+        double[][] test2 = new double[3][3];
         test2[0][0] = 0;
         test2[0][1] = 0;
         test2[0][2] = 2;
